+++ Scribe/Scribe-Api/src/utopia/scribe/api/app/console/ScribeConsoleSettings.scala
@@ -1,20 +1,21 @@
 package utopia.scribe.api.app.console
 
 import utopia.bunnymunch.jawn.JsonBunny
+import utopia.flow.async.context.ThreadPool
 import utopia.flow.collection.CollectionExtensions._
 import utopia.flow.generic.casting.ValueConversions._
-import utopia.flow.generic.factory.FromModelFactoryWithSchema
-import utopia.flow.generic.model.immutable.{Model, ModelDeclaration, PropertyDeclaration}
-import utopia.flow.generic.model.mutable.DataType.StringType
-import utopia.flow.operator.equality.EqualsExtensions._
+import utopia.flow.generic.model.immutable.Model
 import utopia.flow.parse.file.FileExtensions._
-import utopia.flow.parse.file.FileUtils
-import utopia.flow.util.StringExtensions._
+import utopia.flow.parse.json.JsonParser
+import utopia.flow.time.TimeExtensions._
+import utopia.flow.util.AppConfig
+import utopia.flow.util.console.ConsoleExtensions._
 import utopia.flow.util.logging.{Logger, SysErrLogger}
-import utopia.vault.database.Connection
+import utopia.scribe.api.database.access.logging.issue.AccessIssues
+import utopia.vault.database.{Connection, ConnectionPool}
 
-import java.io.FileNotFoundException
 import java.nio.file.Path
+import scala.io.StdIn
 
 /**
   * An interface for user-specified settings (json-based)
@@ -25,70 +26,122 @@ object ScribeConsoleSettings
 {
 	// ATTRIBUTES   ----------------
 	
-	private lazy val loaded = {
-		// Looks recursively for a settings file
-		// Prefers a json file which mentions both "scribe" and "settings",
-		// but if such file is not found, looks for any json file containing the word "settings".
+	/**
+	 * JSON parser used during console sessions
+	 */
+	implicit val jsonParser: JsonParser = JsonBunny
+	/**
+	 * Execution context to use in the console
+	 */
+	implicit val threadPool: ThreadPool = new ThreadPool("Scribe-Console", 3, 100, 10.seconds)(SysErrLogger)
+	/**
+	 * DB connection pool to use in the console.
+	 * Please use this only if [[setupDb]] yields a success.
+	 */
+	implicit val cPool: ConnectionPool = new ConnectionPool(30)
+	
+	/**
+	 * App configuration file access
+	 */
+	private val config = {
 		implicit val log: Logger = SysErrLogger
-		val settingsFileOptions = FileUtils.workingDirectory.toTree.topDownNodesBelowIterator.map { _.nav }
-			.filter { f =>
-				val (fileName, fileType) = f.fileNameAndType.toTuple
-				fileName.containsIgnoreCase("settings") && (fileType ~== "json")
-			}
-			.caching
-		settingsFileOptions.find { _.fileName.containsIgnoreCase("scribe") }.orElse { settingsFileOptions.headOption }
-			.toTry { new FileNotFoundException("No settings.json file found") }
-			// Attempts to parse the file contents, also
-			.flatMap { JsonBunny(_).flatMap { _.tryModel.flatMap(Settings.apply) } }
+		AppConfig("scribe", allowWorkingDirectoryAsAppDirectory = true).logToTry
 	}
 	
 	
 	// COMPUTED --------------------
 	
 	/**
-	  * @return The directory where errors should be logged.
-	  *         None if no directory has been specified.
+	  * @return The directory where errors should be logged. None if no directory has been specified.
 	  */
-	def logDirectory = loaded.toOption.flatMap { _.logDirectory }
+	def logDirectory = config.toOption.flatMap { _("log_directory", "log").string.map { s => s: Path } }
+	/**
+	 * @param directory Directory where file-logging should be performed
+	 * @return Success or failure, depending on whether the app configuration could be updated
+	 */
+	def logDirectory_=(directory: Path) = config.map { _("log_directory") = directory.toJson }
 	
 	/**
 	  * @return Name of the database to connect to when using Scribe features
 	  */
-	def dbName = loaded.get.dbName
+	def dbName = config.flatMap { _("database:name").tryString }
+	/**
+	 * @param dbName Name of the database to use
+	 * @return Success or failure, depending on whether the app configuration could be updated
+	 */
+	def dbName_=(dbName: String) = {
+		Connection.modifySettings { _.copy(defaultDBName = Some(dbName)) }
+		config.map { _("database:name") = dbName }
+	}
 	
 	
 	// OTHER    --------------------
 	
 	/**
-	  * Initializes database connection settings
-	  * @return Success or failure
+	  * Initializes database connection settings and tests database accessibility
+	  * @param allowUserInteraction Whether this application should request DB access settings from the user,
+	 *                             if they haven't been specified
+	 * @return Success containing the name of the targeted database, or a failure.
+	 *         If failure, the database can't be accessed.
 	  */
-	def initializeDbSettings() = loaded.map { settings =>
-		Connection.modifySettings { _.copy(connectionTarget = settings.dbAddress, user = settings.dbUser,
-			password = settings.dbPassword, defaultDBName = Some(settings.dbName))
+	def setupDb(allowUserInteraction: Boolean = false) = {
+		config.flatMap { config =>
+			val defaultConnectionTarget = "jdbc:mariadb://localhost:3306/"
+			// Attempts to load previously specified settings
+			val loadResult = config("database").tryModel.flatMap { db =>
+				db("password").tryString.map { password =>
+					val dbName = db("name").stringOr("scribe_db")
+					Connection.modifySettings { _.copy(
+						connectionTarget = db("address").stringOr(defaultConnectionTarget),
+						user = db("user").stringOr("root"), password = password, defaultDBName = Some(dbName))
+					}
+					dbName
+				}
+			}
+			// Requests settings from the user, if necessary & possible
+			val setupResult = {
+				if (loadResult.isSuccess || !allowUserInteraction ||
+					!StdIn.ask("Database access settings have not been specified. Are you able to specify them now?",
+						default = true))
+					loadResult
+				else {
+					// Requests DB access settings
+					val address = StdIn.readNonEmptyLine(
+						s"Please specify the DB connection address. Default = $defaultConnectionTarget")
+						.getOrElse(defaultConnectionTarget)
+					val user = StdIn.readNonEmptyLine(
+						"Please specify the user to connect to the DB with. Default = root.").getOrElse("root")
+					StdIn.readNonEmptyLine("Please specify the password to access the DB with",
+						"Not specifying a password will cancel this process. Please specify one.")
+						.toTry { new IllegalArgumentException("DB password was not specified") }
+						.map { password =>
+							val dbName = StdIn.readNonEmptyLine(
+								"Please specify the name of the used database. Default = scribe_db.")
+								.getOrElse("scribe_db")
+							
+							// Updates DB access settings
+							Connection.modifySettings { _.copy(
+								connectionTarget = address, user = user, password = password,
+								defaultDBName = Some(dbName))
+							}
+							
+							// Stores the acquired settings
+							config("database") = Model.from("address" -> address, "name" -> dbName,
+								"user" -> user, "password" -> password)
+							
+							dbName
+						}
+				}
+			}
+			setupResult.flatMap { dbName =>
+				// Tests DB access
+				cPool.tryWith { implicit c =>
+					AccessIssues.nonEmpty
+					dbName
+				}
+			}
 		}
 	}
-	
-	
-	// NESTED   --------------------
-	
-	private object Settings extends FromModelFactoryWithSchema[Settings]
-	{
-		override lazy val schema: ModelDeclaration = ModelDeclaration(
-			PropertyDeclaration("log_directory", StringType, Vector("log"), isOptional = true)
-		).withChild("database", ModelDeclaration(
-			PropertyDeclaration("address", StringType, Vector("target", "url"), "jdbc:mariadb://localhost:3306/"),
-			PropertyDeclaration("user", StringType, defaultValue = "root"),
-			PropertyDeclaration("password", StringType, Vector("pw"), isOptional = true),
-			PropertyDeclaration("name", StringType, defaultValue = "scribe_db")
-		))
-		
-		override protected def fromValidatedModel(model: Model): Settings = {
-			val db = model("database").getModel
-			apply(db("address").getString, db("user").getString, db("password").getString, db("name").getString,
-				model("log_directory").string.flatMap { s => (s: Path).asExistingDirectory.toOption })
-		}
-	}
-	private case class Settings(dbAddress: String, dbUser: String, dbPassword: String, dbName: String,
-	                            logDirectory: Option[Path])
+	@deprecated("Renamed to setupDb(Boolean)", "v1.2")
+	def initializeDbSettings(allowUserInteraction: Boolean = false) = setupDb(allowUserInteraction)
 }