+++ Scribe/Scribe-Api/Changes.md
@@ -1,6 +1,10 @@
 # Utopia Scribe Api - list of changes
 
-## v1.1.2 (in development)
+## v1.2 (in development)
+### Breaking changes
+- The Scribe Console settings are no longer read from a JSON file, but from an app configuration file instead 
+  (using **AppConfig**)
+  - Database access settings are requested on the first app startup
 
 ## v1.1.1 - 15.03.2026
 Supports **Flow v2.8**, **Vault v2.1** and **Nexus v2.0**